backend/schemas/user.py
# backend/schemas/user.py

from pydantic import BaseModel

class UserOut(BaseModel):
    id: int
    username: str
    email: str
    role: str
    is_active: bool

    class Config:
        from_attributes = True
