requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
python-dotenv
